Move x86 traits to a module that other architectures can reuse
Eventually, these traits may be exported publicly from the crate
not depending on the architecture.
